Refactor user list notification system

+++ RuntimeLib/Item.c
@@ -400,7 +400,7 @@ RUNTIME_ITEM_PROCEDURE_BINDING(RTItemImmediateReward) {
 		UNIMPLEMENTED;
 		
 	case RUNTIME_ITEM_SUBTYPE_IMMEDIATE_REWARD_MEDALSCORE:
-		if (!RTCharacterAddHonorMedalScore(Runtime, Character, 0, ItemSlot->ItemOptions)) {
+		if (!RTCharacterAddHonorMedalScore(Runtime, Character, 0, (Int32)ItemSlot->ItemOptions)) {
 			return RUNTIME_ITEM_USE_RESULT_FAILED;
 		}
 		break;
@@ -656,7 +656,7 @@ RUNTIME_ITEM_PROCEDURE_BINDING(RTItemSlotConverter) {
 	Int32 SourceItemSlotIndex = ItemSlot->SlotIndex;
 	Int32 TargetItemSlotIndex = TargetItemSlot->SlotIndex;
 
-	Int32 Seed = PlatformGetTickCount();
+	Int32 Seed = (Int32)PlatformGetTickCount();
 	Int32 RandomValue = RandomRange(&Seed, 0, 1000);
 	if (RandomValue < ItemData->SlotConverter.SuccessRate) {
 		TargetItemOptions.Equipment.SlotCount += 1;
@@ -701,7 +701,7 @@ RUNTIME_ITEM_PROCEDURE_BINDING(RTItemEpicConverter) {
 		MaxRandomValue += OptionPool->EpicOptionPoolValueList[Index].Rate;
 	}
 
-	Int32 Seed = PlatformGetTickCount();
+	Int32 Seed = (Int32)PlatformGetTickCount();
 	Int32 RandomValue = RandomRange(&Seed, 0, MaxRandomValue);
 	Int32 RateOffset = 0;
 	Bool Success = false;
@@ -862,7 +862,7 @@ RUNTIME_ITEM_PROCEDURE_BINDING(RTItemStackablePotion) {
 		UNIMPLEMENTED;
 
 	case RUNTIME_ITEM_SUBTYPE_IMMEDIATE_REWARD_MEDALSCORE:
-		if (!RTCharacterAddHonorMedalScore(Runtime, Character, 0, TotalAmount)) {
+		if (!RTCharacterAddHonorMedalScore(Runtime, Character, 0, (Int32)TotalAmount)) {
 			return RUNTIME_ITEM_USE_RESULT_FAILED;
 		}
 		break;
